center map on clicked marker and create javascript modules
