+++ index.html
@@ -15,6 +15,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Antic+Slab&display=swap" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Noto+Serif:ital,wght@0,400;0,700;1,400&display=swap" rel="stylesheet">
 
 </head>
 <body>
@@ -42,6 +43,9 @@
 
     
 
-    <script src="main.js"></script>
+    <script type="module" src="./main.js"></script>
+    <script type="module" src="./src/markers.js"></script>
+    <script type="module" src="./src/country-layers.js"></script>
+
 </body>
 </html>
