+++ src/country-layers.js
@@ -1,0 +1,118 @@
+export function generateMineralProducers(map){
+
+    let hoveredStateId = null;
+
+    map.addSource('nickel-producers', {
+        type: 'geojson',
+        data: './geojson/nickel-producers-countries.geojson'
+    });
+
+    map.addSource('cobalt-producers', {
+        type: 'geojson',
+        data: './geojson/cobalt-producers-countries.geojson'
+    });
+
+    map.addLayer({
+        'id': 'nickel-producers',
+        'type': 'fill',
+        'source': 'nickel-producers',
+        'layout': {},
+        "maxzoom": 3,
+        'paint': {
+            'fill-color': '#5179A0',
+            'fill-opacity': [
+                'case',
+                ['boolean', ['feature-state', 'hover'], false],
+                0.2,
+                0.5
+            ]
+        }
+    });
+
+    map.addLayer({
+        'id': 'cobalt-producers',
+        'type': 'fill',
+        'source': 'cobalt-producers',
+        'layout': {},
+        "maxzoom": 3,
+        'paint': {
+            'fill-color': '#B1CFEC',
+            'fill-opacity': [
+                'case',
+                ['boolean', ['feature-state', 'hover'], false],
+                0.2,
+                0.5
+            ]
+        }
+    });
+
+    map.on('mouseenter', 'nickel-producers', () => {
+        map.getCanvas().style.cursor = 'pointer';
+    });
+
+    // Change it back to a pointer when it leaves.
+    map.on('mouseleave', 'nickel-producers', () => {
+        map.getCanvas().style.cursor = '';
+    });
+
+    map.on('mouseenter', 'cobalt-producers', () => {
+        map.getCanvas().style.cursor = 'pointer';
+    });
+
+    // Change it back to a pointer when it leaves.
+    map.on('mouseleave', 'cobalt-producers', () => {
+        map.getCanvas().style.cursor = '';
+    });
+
+    // When the user moves their mouse over the state-fill layer, we'll update the
+    // feature state for the feature under the mouse.
+
+    function onMouseHoverCountry(sourceId) {
+        map.on('mousemove', sourceId, (e) => {
+
+            function fillStateFunction(e, geoJsonName) {
+                if (e.features.length > 0) {
+                    if (hoveredStateId !== null) {
+                        map.setFeatureState(
+                            { source: geoJsonName, id: hoveredStateId },
+                            { hover: false }
+                        );
+                    }
+                    hoveredStateId = e.features[0].id;
+                    map.setFeatureState(
+                        { source: geoJsonName, id: hoveredStateId },
+                        { hover: true }
+                    );
+                }
+            }
+    
+            fillStateFunction(e, sourceId);
+        });
+    }
+    
+    onMouseHoverCountry('cobalt-producers')
+    onMouseHoverCountry('nickel-producers')
+
+    // When the mouse leaves the state-fill layer, update the feature state of the
+    // previously hovered feature.
+
+    function offMouseHoverCountry(sourceId) {
+        map.on('mouseleave', sourceId, () => {
+
+            function emptyStateFunction(geoJsonName) {
+                if (hoveredStateId !== null) {
+                    map.setFeatureState(
+                        { source: geoJsonName, id: hoveredStateId },
+                        { hover: false }
+                    );
+                }
+                hoveredStateId = null;
+            }
+    
+            emptyStateFunction(sourceId)
+        });
+    }
+
+    offMouseHoverCountry('nickel-producers')
+    offMouseHoverCountry('cobalt-producers')
+}
