+++ style.css
@@ -5,7 +5,8 @@ body {
 
 a {
     color: rgb(0, 0, 0);
-    font-family: 'Antic Slab', serif;
+    font-family: 'Noto Serif', serif;
+    font-size: 14px;
     padding: 5px;
 }
 
@@ -35,10 +36,10 @@ a {
   }
   
 .openbtn {
-    background-image: url(./img/right-arrow.png);
+    background-image: url(./img/rosa-arrow.png);
     background-size: cover;
-    width: 40px;
-    height: 40px;
+    width: 35px;
+    height: 35px;
     cursor: pointer;
     position:absolute; 
     top:10px; 
@@ -46,11 +47,6 @@ a {
     z-index:1;
     transition: all .5s;
   }
-
-.openbtn:hover {
-    transition: all .2s;
-    transform: translateX(+5px);
-}
 
 /* ------------------------------------ */  
 
@@ -92,7 +88,7 @@ a {
 }
 
 .aluminium.mineral-info-box {
-    background-color: #ECCA85;
+    background-color: #c5b6d8;
 }
 
 /* Markers ---------- */
@@ -105,11 +101,11 @@ a {
     cursor: pointer;
 }
 
-.marker:hover{
+/* .marker:hover{
     width: 60px;
     height: 60px;
     transition: all .2s;
-}
+} */
 
 
 .marker.goro {
@@ -133,7 +129,7 @@ a {
   
 .mapboxgl-popup-content {
     text-align: center;
-    font-family: 'Open Sans', sans-serif;
+    font-family: 'Noto Serif', serif;
 }
 
 /* ------------------------------------ */
